+++ model/ReadOnlyModel.java
@@ -2,11 +2,14 @@ package model;
 
 import java.util.ArrayList;
 
+import model.geometry.ReadOnlyPolygon2D;
 import model.geometry.Vector2D;
 import model.softbody.ReadOnlySoftBody;
 
 public interface ReadOnlyModel {
     ArrayList<ReadOnlySoftBody> getReadOnlySoftBodies();
+
+    ArrayList<ReadOnlyPolygon2D> getReadOnlyPolygons();
 
     public int getId();