+++ model/geometry/ReadOnlyPolygon2D.java
@@ -1,0 +1,9 @@
+package model.geometry;
+
+public interface ReadOnlyPolygon2D {
+    public int[] getXArr();
+
+    public int[] getYArr();
+
+    public int getNumPoints();
+}